Feat: 미니쉘 프로젝트 구조 변경
1. 각 파일의 성격에 맞게 디렉토리 구성

resolved #121
