--- srcs/error_handling.c
@@ -1,50 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   error_handling.c                                   :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: sehjang <sehjang@student.42seoul.kr>       +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/09/16 18:56:33 by sehjang           #+#    #+#             */
-/*   Updated: 2022/09/16 18:56:33 by sehjang          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../includes/minishell.h"
-
-void	print_error_message_with_token(char *cmd, char *token, char *msg)
-{
-	ft_putstr_fd("minishell: ", 2);
-	ft_putstr_fd(cmd, 2);
-	ft_putstr_fd(": ", 2);
-	ft_putstr_fd(token, 2);
-	ft_putstr_fd(": ", 2);
-	ft_putstr_fd(msg, 2);
-	ft_putstr_fd("\n", 2);
-}
-
-void	print_error_message(char *cmd, char *msg)
-{
-	ft_putstr_fd("minishell: ", 2);
-	ft_putstr_fd(cmd, 2);
-	ft_putstr_fd(": ", 2);
-	ft_putstr_fd(msg, 2);
-	ft_putstr_fd("\n", 2);
-}
-
-int	print_error_message_syntax(char *token)
-{
-	ft_putstr_fd("minishell: syntax error near unexpected token '", 2);
-	ft_putstr_fd(token, 2);
-	ft_putstr_fd("'\n", 2);
-	return (1);
-}
-
-void	error_excute(char *cmd, char *token, char *msg, int exit_code)
-{
-	if (token != 0)
-		print_error_message_with_token(cmd, token, msg);
-	else
-		print_error_message(cmd, msg);
-	exit(exit_code);
-}
